# docs/contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "additionalProperties": {},
        "description": "D\u1EEF li\u1EC7u l\u1ED7i chu\u1EA9n tr\u1EA3 v\u1EC1 t\u1EEB ki\u1EC3m tra d\u1EEF li\u1EC7u, quy t\u1EAFc nghi\u1EC7p v\u1EE5 v\u00E0 l\u1EDBp x\u1EED l\u00FD ngo\u1EA1i l\u1EC7.",
        "properties": {
          "code": {
            "description": "M\u00E3 l\u1ED7i \u1ED5n \u0111\u1ECBnh d\u00E0nh cho \u1EE9ng d\u1EE5ng g\u1ECDi API.",
            "nullable": true,
            "type": "string"
          },
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "details": {
            "description": "Th\u00F4ng tin ch\u1EA9n \u0111o\u00E1n ch\u1EC9 xu\u1EA5t hi\u1EC7n trong m\u00F4i tr\u01B0\u1EDDng ph\u00E1t tri\u1EC3n khi c\u00F3.",
            "nullable": true,
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "L\u1ED7i ki\u1EC3m tra d\u1EEF li\u1EC7u theo t\u00EAn thu\u1ED9c t\u00EDnh c\u1EE7a y\u00EAu c\u1EA7u.",
            "nullable": true,
            "type": "object"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "message": {
            "description": "Th\u00F4ng b\u00E1o l\u1ED7i d\u00E0nh cho ng\u01B0\u1EDDi d\u00F9ng.",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "traceId": {
            "description": "M\u00E3 truy v\u1EBFt y\u00EAu c\u1EA7u d\u00F9ng \u0111\u1EC3 \u0111i\u1EC1u tra s\u1EF1 c\u1ED1.",
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CategoryResponse": {
        "additionalProperties": false,
        "properties": {
          "children": {
            "items": {
              "$ref": "#/components/schemas/CategoryResponse"
            },
            "nullable": true,
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "nullable": true,
            "type": "string"
          },
          "parentId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "parentName": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateCategoryRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "nullable": true,
            "type": "string"
          },
          "parentId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "MessageResponse": {
        "additionalProperties": false,
        "description": "D\u1EEF li\u1EC7u th\u00F4ng b\u00E1o th\u00E0nh c\u00F4ng.",
        "properties": {
          "message": {
            "description": "Th\u00F4ng b\u00E1o th\u00E0nh c\u00F4ng d\u00E0nh cho ng\u01B0\u1EDDi d\u00F9ng.",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateCategoryRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "nullable": true,
            "type": "string"
          },
          "parentId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "description": "Nh\u1EADp token d\u1EA1ng: Bearer {token}",
        "scheme": "Bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API E-Commerce phi\u00EAn b\u1EA3n 1. Route /api/v1 \u0111\u01B0\u1EE3c khuy\u1EBFn ngh\u1ECB; route /api \u0111\u01B0\u1EE3c gi\u1EEF \u0111\u1EC3 t\u01B0\u01A1ng th\u00EDch ng\u01B0\u1EE3c.",
    "title": "ECommerce API",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Y\u00EAu c\u1EA7u kh\u00F4ng h\u1EE3p l\u1EC7 do d\u1EEF li\u1EC7u \u0111\u1EA7u v\u00E0o ho\u1EB7c quy t\u1EAFc nghi\u1EC7p v\u1EE5."
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "L\u1ED7i x\u1EED l\u00FD n\u1ED9i b\u1ED9 c\u1EE7a m\u00E1y ch\u1EE7."
          }
        },
        "summary": "L\u1EA5y t\u1EA5t c\u1EA3 danh m\u1EE5c (d\u1EA1ng c\u00E2y cha-con)",
        "tags": [
          "Category"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Y\u00EAu c\u1EA7u kh\u00F4ng h\u1EE3p l\u1EC7 do d\u1EEF li\u1EC7u \u0111\u1EA7u v\u00E0o ho\u1EB7c quy t\u1EAFc nghi\u1EC7p v\u1EE5."
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Ch\u01B0a x\u00E1c th\u1EF1c - m\u00E3 truy c\u1EADp b\u1ECB thi\u1EBFu ho\u1EB7c kh\u00F4ng h\u1EE3p l\u1EC7."
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Ng\u01B0\u1EDDi d\u00F9ng \u0111\u00E3 x\u00E1c th\u1EF1c nh\u01B0ng kh\u00F4ng c\u00F3 \u0111\u1EE7 quy\u1EC1n th\u1EF1c hi\u1EC7n thao t\u00E1c."
          },
          "409": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "D\u1EEF li\u1EC7u xung \u0111\u1ED9t do t\u00E0i nguy\u00EAn \u0111\u00E3 \u0111\u01B0\u1EE3c m\u1ED9t thao t\u00E1c kh\u00E1c thay \u0111\u1ED5i."
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "L\u1ED7i x\u1EED l\u00FD n\u1ED9i b\u1ED9 c\u1EE7a m\u00E1y ch\u1EE7."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "[Admin] T\u1EA1o danh m\u1EE5c m\u1EDBi",
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Y\u00EAu c\u1EA7u kh\u00F4ng h\u1EE3p l\u1EC7 do d\u1EEF li\u1EC7u \u0111\u1EA7u v\u00E0o ho\u1EB7c quy t\u1EAFc nghi\u1EC7p v\u1EE5."
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Ch\u01B0a x\u00E1c th\u1EF1c - m\u00E3 truy c\u1EADp b\u1ECB thi\u1EBFu ho\u1EB7c kh\u00F4ng h\u1EE3p l\u1EC7."
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Ng\u01B0\u1EDDi d\u00F9ng \u0111\u00E3 x\u00E1c th\u1EF1c nh\u01B0ng kh\u00F4ng c\u00F3 \u0111\u1EE7 quy\u1EC1n th\u1EF1c hi\u1EC7n thao t\u00E1c."
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Kh\u00F4ng t\u00ECm th\u1EA5y t\u00E0i nguy\u00EAn \u0111\u01B0\u1EE3c y\u00EAu c\u1EA7u."
          },
          "409": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "D\u1EEF li\u1EC7u xung \u0111\u1ED9t do t\u00E0i nguy\u00EAn \u0111\u00E3 \u0111\u01B0\u1EE3c m\u1ED9t thao t\u00E1c kh\u00E1c thay \u0111\u1ED5i."
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "L\u1ED7i x\u1EED l\u00FD n\u1ED9i b\u1ED9 c\u1EE7a m\u00E1y ch\u1EE7."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "[Admin] X\u00F3a m\u1EC1m danh m\u1EE5c",
        "tags": [
          "Category"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Y\u00EAu c\u1EA7u kh\u00F4ng h\u1EE3p l\u1EC7 do d\u1EEF li\u1EC7u \u0111\u1EA7u v\u00E0o ho\u1EB7c quy t\u1EAFc nghi\u1EC7p v\u1EE5."
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Kh\u00F4ng t\u00ECm th\u1EA5y t\u00E0i nguy\u00EAn \u0111\u01B0\u1EE3c y\u00EAu c\u1EA7u."
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "L\u1ED7i x\u1EED l\u00FD n\u1ED9i b\u1ED9 c\u1EE7a m\u00E1y ch\u1EE7."
          }
        },
        "summary": "L\u1EA5y chi ti\u1EBFt danh m\u1EE5c theo Id",
        "tags": [
          "Category"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Y\u00EAu c\u1EA7u kh\u00F4ng h\u1EE3p l\u1EC7 do d\u1EEF li\u1EC7u \u0111\u1EA7u v\u00E0o ho\u1EB7c quy t\u1EAFc nghi\u1EC7p v\u1EE5."
          },
          "401": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Ch\u01B0a x\u00E1c th\u1EF1c - m\u00E3 truy c\u1EADp b\u1ECB thi\u1EBFu ho\u1EB7c kh\u00F4ng h\u1EE3p l\u1EC7."
          },
          "403": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Ng\u01B0\u1EDDi d\u00F9ng \u0111\u00E3 x\u00E1c th\u1EF1c nh\u01B0ng kh\u00F4ng c\u00F3 \u0111\u1EE7 quy\u1EC1n th\u1EF1c hi\u1EC7n thao t\u00E1c."
          },
          "404": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Kh\u00F4ng t\u00ECm th\u1EA5y t\u00E0i nguy\u00EAn \u0111\u01B0\u1EE3c y\u00EAu c\u1EA7u."
          },
          "409": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "D\u1EEF li\u1EC7u xung \u0111\u1ED9t do t\u00E0i nguy\u00EAn \u0111\u00E3 \u0111\u01B0\u1EE3c m\u1ED9t thao t\u00E1c kh\u00E1c thay \u0111\u1ED5i."
          },
          "500": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "L\u1ED7i x\u1EED l\u00FD n\u1ED9i b\u1ED9 c\u1EE7a m\u00E1y ch\u1EE7."
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "[Admin] C\u1EADp nh\u1EADt danh m\u1EE5c",
        "tags": [
          "Category"
        ]
      }
    },
...
  }
}